pyproject.toml
[project]
name = "freight-brokerage-abm"
version = "0.1.0"
description = "Agent-based model simulating freight brokerage negotiations"
authors = [
    {name = "Agent Based Modelling Project"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "mesa (>=3.2.0,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
